requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.95.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
